% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudio-package.R
\docType{data}
\name{alt}
\alias{alt}
\title{Sonoran desert altitude.}
\description{
This is a raster file for altitude in the Sonoran desert
 region coincident with the Lophocereus and Upiga data
 sets.
}
\keyword{data}
