\docType{data}
\name{araptus_attenuatus}
\alias{araptus_attenuatus}
\title{An example data set for co-dominant loci}
\description{
  This data set was published in Garrick et al. (2012) and
  consists of four strata, two coordinates, and eight
  co-dominant loci.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}
\keyword{data}
\keyword{dataset}

