% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.data.frame.genind}
\alias{as.data.frame.genind}
\title{Converts genind to normal data.frame with \code{locus} objects}
\usage{
as.data.frame.genind(x)
}
\arguments{
\item{x}{An object of type \code{genind}}
}
\value{
A \code{data.frame} object with loci represented as \code{locus} objects.
}
\description{
This is a conversion function to translate an object of type \code{genind} from
 the adegenet package into a normal data frame with \code{locus} objects.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

