% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudio-package.R
\docType{data}
\name{baja}
\alias{baja}
\title{Metadata for Baja Populations.}
\description{
This is metadata associated with the sampling locations
 for the Lophocereus and Upiga data sets.
}
\keyword{data}
