% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_distance.R
\name{centroid_distance}
\alias{centroid_distance}
\title{This function takes a data set and a group set and returns centroid distances}
\usage{
centroid_distance(x, grouping)
}
\arguments{
\item{x}{The NxP matrix of raw data.}

\item{grouping}{The Nx1 vector of grouping factors.}
}
\value{
An NxN pairwise distance matrix.
}
\description{
This is a convience function for a reduce and distance operation based upon
 multivariate predictor variables.
}
\examples{
data(arapat)
mv <- to_mv(arapat)
cd <- centroid_distance(mv, arapat$Population)
dim(cd)
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
