% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_variance.R
\name{centroid_variance}
\alias{centroid_variance}
\title{This function takes a data set and a group set and returns centroid variances}
\usage{
centroid_variance(x, grouping)
}
\arguments{
\item{x}{The NxP matrix of raw data.}

\item{grouping}{The Nx1 vector of grouping factors.}
}
\value{
A \code{data.frame} with one row per group: \code{Stratum} and
  \code{variance}, the sum of the within-stratum variances.
}
\description{
This is a convience function for a reduce and distance operation based upon
 multivariate predictor variables.
}
\examples{
data(arapat)
mv <- to_mv(arapat)
cv <- centroid_variance(mv, arapat$Population)
head(cv)
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
