\docType{data}
\name{cornus_florida}
\alias{cornus_florida}
\title{An example data set for parent offspring data}
\description{
  This data set was published in Dyer et al. (2012) and
  consists of 22 adult and 41 offspring from two mothers.
  Genotypes are assayed for
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}
\keyword{data}
\keyword{dataset}

