% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_resistances.R
\name{create_resistances}
\alias{create_resistances}
\title{Estimates resistance rasters from binary feature raster.}
\usage{
create_resistances(x, feature_name = "layer", costs = c(2, 5, 10, 50,
  100), symmetric = TRUE)
}
\arguments{
\item{x}{A base \code{raster} object from which to base others.}

\item{feature_name}{The prefix for output files for this raster feature (default='layer').}

\item{costs}{A vector of cost rasters to iterate across (default=c(2, 5, 10, 50, 100)).}

\item{symmetric}{A flag indicating that costs out of as well as into features will be estimated.}
}
\description{
This function produces a cost raster from a single value
}
\examples{
vals <- sample( c(0,1), size=49, replace=TRUE)
m <- matrix( vals, nrow=7)
r <- raster::raster( m )
create_resistances( r )

}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
