% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2cov.R
\name{dist2cov}
\alias{dist2cov}
\title{Converts distance matrix to covarance one.}
\usage{
dist2cov(D)
}
\arguments{
\item{D}{A pairwise, square, and symmetric distance matrix.}
}
\value{
A covariance matrix of the same size.
}
\description{
This function takes a distance matrix and turns it into a covariance
  matrix following Gower (1966).
}
\examples{
D <- matrix(c(0, 1, 1, 0), nrow = 2)
dist2cov(D)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
