% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_covariance.R
\name{distance_to_covariance}
\alias{distance_to_covariance}
\title{Takes distance matrix and turns it into a covariance matrix}
\usage{
distance_to_covariance(X)
}
\arguments{
\item{X}{A square and symmetric distance matrix.}
}
\value{
A covariance matrix the same size as \code{X}
}
\description{
This function takes a distance matrix and turns it into a 
  covariance matrix.
}
\examples{
D <- matrix( c(0,7,6,7,7,0,13,8,6,13,0,4,7,8,4,0), nrow=4)
C <- distance_to_covariance(D)
C
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

