% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pop.R
\name{.get_migration_matrix}
\alias{.get_migration_matrix}
\title{Look up the active migration matrix for a given generation}
\usage{
.get_migration_matrix(gen, events)
}
\arguments{
\item{gen}{Current generation number.}

\item{events}{List of migration_event objects (or NULL).}
}
\value{
A migration matrix or NULL.
}
\description{
Look up the active migration matrix for a given generation
}
\keyword{internal}
