% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_loci.R
\name{.mutate_locus_vector}
\alias{.mutate_locus_vector}
\title{Mutate a vector of locus objects}
\usage{
.mutate_locus_vector(locus_vec, mutation)
}
\arguments{
\item{locus_vec}{A vector of \code{locus} objects.}

\item{mutation}{A \code{mutation_model} object.}
}
\value{
A vector of \code{locus} objects.
}
\description{
Mutate a vector of locus objects
}
\keyword{internal}
