% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_pop.R
\name{.save_census}
\alias{.save_census}
\title{Save a census snapshot as an .rds file}
\usage{
.save_census(pop, gen, census_dir)
}
\arguments{
\item{pop}{data.frame of the population.}

\item{gen}{Generation number.}

\item{census_dir}{Directory to write to.}
}
\description{
Save a census snapshot as an .rds file
}
\keyword{internal}
