% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_loci.R
\name{.zero_pad}
\alias{.zero_pad}
\title{Zero-pad a numeric value to match original allele width}
\usage{
.zero_pad(val, width)
}
\arguments{
\item{val}{Numeric value.}

\item{width}{Target character width.}
}
\value{
Character string, zero-padded.
}
\description{
Zero-pad a numeric value to match original allele width
}
\keyword{internal}
