\name{exclusion.probability}
\alias{exclusion.probability}
\title{Estimates the exclusion probability from a Frequencies object}
\usage{
  exclusion.probability(freq)
}
\arguments{
  \item{freq}{A Frequencies object calculated from pollen
  pool allele frequencies or from a set of adults.}
}
\value{
  Single locus exclusion probability
}
\description{
  Estimates the exclusion probability from a Frequencies
  object
}
\note{
  This function estiamtes the single-locus exclusion
  probability.  It is left up to you to decide what group
  of individuals you are going to use to estimate
  background allele frequencies.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

