% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fis.R
\name{Fis}
\alias{Fis}
\title{Estimate simple inbreeding from frequencies}
\usage{
Fis(x, small.N = FALSE, stratum = NULL, loci = NULL)
}
\arguments{
\item{x}{Either a \code{locus} object or a data.frame with locus objects.}

\item{small.N}{Passes this along to He for small sample sizes.}

\item{stratum}{An optional term for the estimation of inbreeding from samples of populations 
(default=NULL)}

\item{loci}{The subset of loci to use (default = all)}
}
\value{
The inbreeding F statistic as a \code{numeric} value or a \code{data.frame}
 if you passed multiple loci to this function.
}
\description{
Returns the general Fis = 1-ho/he parameter from the locus being passed or
 a set of them if the value passed is a \code{data.frame} with locus objects.
}
\examples{
loci <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B") ) )
Fis( loci )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

