\docType{class}
\name{Frequencies-class}
\alias{Frequencies-class}
\title{Class \code{Frequencies}}
\description{
  The \code{Frequencies} class is the base class that
  estimates allele frequencies for a single locus.  It can
  also provide estimates of observed and expected
  heterozygosity.
}
\note{
  Notes on \code{Frequencies}
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu
}
\keyword{classes}

