% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read012.R
\name{from012}
\alias{from012}
\title{Imports from 012 files}
\usage{
from012(fileName)
}
\arguments{
\item{fileName}{The name of the base file with the 012 encoding. This is
typically something like 'out.012'.  This function will infer that there 
are 2 more files in the same location, one ending in `.indiv` that will be 
the names of the locales where the individuals were sampled, and `.pos` 
which will indicate the contig onto which the}
}
\value{
A data.frame with formatted genotypes
}
\description{
This function reads in data that is stored in 012 format into gstudio
 and
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
