% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{Fst}
\alias{Fst}
\title{Estimation Wright's Fst parameter}
\usage{
Fst(x, stratum = "Population", nperm = 0)
}
\arguments{
\item{x}{A \code{data.frame} with \code{locus} objects.  If there is more than one 
locus present, it will estimate the parameter for each locus.  It DOES NOT estimate
the multilocus Fst (and do not even think of averaging the single locus estimates).}

\item{stratum}{Either the name of the column representing strata in \code{x}.  By 
default, this function will use "Population".}

\item{nperm}{The number of permutations to run to test Fst=0 hypothesis.}
}
\value{
An \code{data.frame} with Fst, sigma_p (variance among populations), and pq
 the total variance at the locus.
}
\description{
This function estimates Wright's Fst parameter, the one based upon the variance
 in p, most applicable to a 2-allele locus.  The results are returned
 as a \code{data.frame}.
}
\examples{
 AA <- locus( c("A","A") )
 AB <- locus( c("A","B") )
 BB <- locus( c("B","B") )
 locus <- c(AA,AA,AA,AA,BB,BB,BB,AB,AB,AA)
 Population <- c(rep("Pop-A",5),rep("Pop-B",5))
 df <- data.frame(Population,locus)
 Fst( df )
 locus2 <- c(AA,AA,AA,AA,AA,BB,BB,BB,BB,BB)
 df <- data.frame( Population, TPI=locus, PGM=locus2 )
 Fst( df )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

