% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_relatedness.R
\name{genetic_relatedness}
\alias{genetic_relatedness}
\title{Estimates pair-wise relatedness}
\usage{
genetic_relatedness(
  x,
  loci = NA,
  mode = c("Nason", "LynchRitland", "Ritland", "Queller")[1],
  freqs = NA
)
}
\arguments{
\item{x}{A \code{data.frame} with \code{locus} columns, or a single \code{locus} vector.}

\item{loci}{The loci to use (if omitted, all loci are used).}

\item{mode}{The relatedness metric to use:
\describe{
  \item{Nason}{Fij estimator of coancestry (Nason)}
  \item{LynchRitland}{Lynch & Ritland (1999) regression estimator}
  \item{Ritland}{Ritland (1996) estimator}
  \item{Queller}{Queller & Goodnight (1989) allele-wise relatedness}
}}

\item{freqs}{An optional \code{data.frame} of allele frequencies as returned by \code{\link{frequencies}}.
If omitted, estimated from the data.}
}
\value{
A matrix of pairwise relatedness estimates.
}
\description{
This is the primary front-end function for estimating pairwise genetic relatedness
 among individuals.
}
\note{
This function operates on diploid data and will return NA for any 
 comparison of missing genotypes.
}
\examples{
loc1 <- c( locus(c("A","A")), locus(c("A","B")), locus(c("B","B")) )
loc2 <- c( locus(c("1","1")), locus(c("1","2")), locus(c("2","2")) )
df <- data.frame( TPI=loc1, PGM=loc2 )
genetic_relatedness( df, mode="Nason" )
genetic_relatedness( df, mode="LynchRitland" )
genetic_relatedness( df, mode="Ritland" )
genetic_relatedness( df, mode="Queller" )
}
