% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_counts.R
\name{genotype_counts}
\alias{genotype_counts}
\title{Returns the number of non-missing genotypes per stratum}
\usage{
genotype_counts(x, stratum = NULL)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{stratum}{The name of the column to use as a stratum.  If not given, the entire
data set is used (default=NULL which will lump all data).}
}
\value{
A \code{data.frame} with Stratum (set to "All" if not passed as param) and the
 number of observations without missing genetic data per column.
}
\description{
This function provides a quick summary of the observed data, by stratum (optional)
 of the number of samples without missing data.
}
\examples{
data(arapat)
gc <- genotype_counts(arapat, stratum = "Population")
head(gc)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
