\name{google.pie.chart}
\alias{google.pie.chart}
\title{Get Google Pie Chart URL}
\usage{
  google.pie.chart(x, color, labels = NULL)
}
\arguments{
  \item{x}{The values to create the pie chart from}

  \item{labels}{The labels to attach to those values}

  \item{color}{The colors to use.}
}
\value{
  A URL
}
\description{
  This function creates a fully formed URL for extracting
  pie charts from the google charts static API.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}
\seealso{
  \code{pies.on.map}
}

