\docType{package}
\name{gstudio-package}
\alias{gstudio-package}
\title{Spatial analysis of genetic marker data}
\description{
  gstudio is a suite of classes and functions for the
  spatial analysis of genetic marker data.
}
\details{
  \tabular{ll}{ Package: \tab gstudio\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2011-08-11\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

  There are some very good examples of how \code{Locus} and
  \code{Population} classes are used along with
  \code{population.graph} approaches located in the
  vignettes; \file{DataImport.pdf},
  \file{GeneticSummaries.pdf}, \file{GeneticDistance.pdf},
  \file{GeneticStructure}, and \file{PopulationGraph.pdf}.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}
\keyword{package}

