% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_mean.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{This is a quick and dirty harmonic mean calculation}
\usage{
harmonic_mean(x)
}
\arguments{
\item{x}{a vector of values}
}
\value{
The harmonic mean
}
\description{
This estimates the harmonic mean.
}
\examples{
 x <- c(1,2,3,4,5)
 harmonic_mean(x)
 x <- c(1,2,3,4,5,0)
 harmonic_mean(x)
}

