% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he.R
\name{He}
\alias{He}
\title{Estimate expected heterozygosity
 
Returns the general expected heterozygosity parameter
 from the frequencies}
\usage{
He(x, small.N = FALSE, stratum = NULL)
}
\arguments{
\item{x}{Either a \code{data.frame} object with \code{locus} objects or 
a vector or \code{locus} objects.}

\item{small.N}{Apply the 2N/(2N-1) correction to the data
for small sample sizes.}

\item{stratum}{This optional argument makes He estimate Nei's Hs parameter taking
into consideration population subdivision.  This is an unbiased estimator.}
}
\value{
The expected heterozygosity as a numeric or a \code{data.frame} if 
 several loci are passed.
}
\description{
Estimate expected heterozygosity
 
Returns the general expected heterozygosity parameter
 from the frequencies
}
\examples{
loci <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B") ) )
He( loci )
He( loci, small.N=TRUE )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

