% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hes.R
\name{Hes}
\alias{Hes}
\title{Estimate expected heterozygosity
 
Returns the general expected heterozygosity parameter
 from the frequencies}
\usage{
Hes(x, stratum = "Population", small.N = FALSE)
}
\arguments{
\item{x}{Either a \code{data.frame} object with \code{locus} objects or 
a vector or \code{locus} objects.}

\item{stratum}{This optional argument makes He estimate Nei's Hs parameter taking
into consideration population subdivision.  This is an unbiased estimator.}

\item{small.N}{Apply the 2N/(2N-1) correction to the data
for small sample sizes.}
}
\value{
The expected heterozygosity as a numeric or a \code{data.frame} if 
 several loci are passed.
}
\description{
Estimate expected heterozygosity
 
Returns the general expected heterozygosity parameter
 from the frequencies
}
\examples{
data(arapat)
Hes( arapat )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
