% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroscedasticity.R
\name{het_data}
\alias{het_data}
\title{Heteroscedasticity}
\usage{
het_data()
}
\description{
This function takes a set of multilocus genotypes and estimates a vector divergence from stratum centroids.  It
 can then
}
\examples{
library( dplyr)  
data(arapat)
arapat |> 
 filter( Species == "Mainland" ) |> 
 select( -Species, -Cluster, -ID, -Latitude, -Longitude ) |>
 heteroscedasticity()
}
