% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{Hi}
\alias{Hi}
\title{Estimate Individual Heterozygosity
 
Returns the heterozygosity of loci within each individual.  This is a per-individual estiamte}
\usage{
Hi(x)
}
\arguments{
\item{x}{A \code{data.frame} that has rows of individuals who have one or more \code{locus} 
columns}
}
\value{
The average heterozygosity for non-missing loci per individual as a \code{data.frame} 
 object with all non-locus columns identical to those passed.
}
\description{
Estimate Individual Heterozygosity
 
Returns the heterozygosity of loci within each individual.  This is a per-individual estiamte
}
\note{
This function must be called on a \code{data.frame} of individuals, it will assume that
 each row contains at least one column of type \code{locus}
}
\examples{
loc1 <- c( locus( c("A","A") ), locus( c("A","B") ), locus( c("A","B") ) )
loc2 <- c( locus( c("A","B") ), locus( c("A","A") ), locus( c("A","B") ) )
loc3 <- c( locus( c("A","B") ), locus( c("A","A") ), locus( c("A","B") ) )
loc4 <- c( locus( c("A","A") ), locus( c("A","A") ), NA )
pop <- c("A","A","B")
df <- data.frame( ID = 1:3, pop, loc1, loc2, loc3, loc4 )
Hi( df )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
