% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho.R
\name{Ho}
\alias{Ho}
\title{Estimate observed heterozygosity
 
Returns the general observed heterozygosity parameter
 from the genotypes.}
\usage{
Ho(x)
}
\arguments{
\item{x}{An object of type \code{locus} or a \code{data.frame} with locus columns.}
}
\value{
The observed heterozygosity.
}
\description{
Estimate observed heterozygosity
 
Returns the general observed heterozygosity parameter
 from the genotypes.
}
\note{
For stratified estimates across populations, use \code{\link{Hos}}.
}
\examples{
loci <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B") ) )
Ho( loci )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
