% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hos.R
\name{Hos}
\alias{Hos}
\title{Estimate observed heterozygosity corrected for sample size after Nei
 
Returns the general observed heterozygosity parameter
 from the frequencies}
\usage{
Hos(x, stratum = "Population")
}
\arguments{
\item{x}{An object of type \code{locus}}

\item{stratum}{A flag indicating that you want to estimate Nei's unbiased Ho
across sampling locations.}
}
\value{
The expected heterozygosity
}
\description{
Estimate observed heterozygosity corrected for sample size after Nei
 
Returns the general observed heterozygosity parameter
 from the frequencies
}
\note{
This function can be called on a single vector of data, a \code{data.frame} of loci, 
 or a \code{data.frame} of \code{locus} objects across strata.  If the estimating across
 stratum, the unbiased estimator should be used to average across stratum and is performed
 by passing the appropriate stratum= argument.
}
\examples{
data(arapat)
Hos(arapat)
loci <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B") ) )
Hos( loci )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

