% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{Ht}
\alias{Ht}
\title{Estimate expected heterozygosity
 
Returns the general expected total heterozygosity parameter}
\usage{
Ht(x, stratum = "Population")
}
\arguments{
\item{x}{A \code{data.frame} object with \code{locus} objects}

\item{stratum}{The name of the column representing the stratum variable (default=Population)}
}
\value{
A \code{data.frame} with columns \code{Locus} and \code{Ht}.
}
\description{
Estimate expected heterozygosity
 
Returns the general expected total heterozygosity parameter
}
\examples{
loci1 <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B")))
loci2 <- c( locus( c("A","A") ), locus( c("A","B")), locus(c("B","B")))
df <- data.frame( Population=c("One","One","One","Two","Two","Two"), Locus=c(loci1,loci2) )
Ht( df )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
