\docType{methods}
\name{is.heterozygote}
\alias{is.heterozygote,Locus-method}
\alias{is.heterozygote}
\title{Is a locus a heterozygote}
\arguments{
  \item{locus}{A \code{Locus} object}
}
\description{
  Is a locus a heterozygote
}
\examples{
\dontrun{

loc <- Locus( c("A","T") )
is.heterozygote(loc)

}
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

