% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_na.R
\name{is.na.locus}
\alias{is.na.locus}
\title{Determines if \code{locus} is NA}
\usage{
\method{is.na}{locus}(x)
}
\arguments{
\item{x}{The \code{locus} object}
}
\value{
A logical flag indicating if \code{x} is NA
}
\description{
Convienence function to determine if the \code{locus} object is NA
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

