% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_frequency.R
\name{is_frequency}
\alias{is_frequency}
\title{Is-A function for Allele Frequencies}
\usage{
is_frequency(x)
}
\arguments{
\item{x}{A \code{data.frame} potentially from \code{frequencies}
function.}
}
\value{
A flag indicating it has the qualities of a \code{data.frame}
 representing allele frequencies
}
\description{
This is a function that returns a flag indicating that the 
 object passed has the qualities of a \code{data.frame} created
 by the function \code{frequencies}.
}
\examples{
df <- data.frame(Allele = c("A", "B"), Frequency = c(0.5, 0.5))
is_frequency(df)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
