% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_loci.R
\name{make_loci}
\alias{make_loci}
\title{Make loci}
\usage{
make_loci(x, N = 20, F = 0)
}
\arguments{
\item{x}{A \code{data.frame} with Allele and Frequency columns.  The Locus column specifies
the names of the alleles to be used and the Frequency one gives their occurrence.}

\item{N}{The number of loci to return (default=20)}

\item{F}{A numeric value for the inbeeding statistic (default=0)}
}
\value{
A vector of \code{locus} objects.
}
\description{
This function makes a vector of locus objects given a frequency array and an 
 estimate of the inbreeding frequency. Given that it is not possible to 'perfectly' 
 make the right distribution of genotypes for all given N & F combinations, this will
 work to provide as close of an approximation as possible.
}
\examples{
 x <- data.frame(Allele=c("A","B"), Frequency=c(0.5,0.5))
 make_loci(x)
 make_loci(x,F=1)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

