% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_populations.R
\name{make_populations}
\alias{make_populations}
\title{Make Populations}
\usage{
make_populations(freqs, N, F = 0)
}
\arguments{
\item{freqs}{The data.frame with Population, Locus, Allele, and Frequency columns}

\item{N}{The number of individuals to make in each population}

\item{F}{The inbreeding level for the individuals}
}
\value{
A data.frame of fully formed individuals
}
\description{
Creates a series of populaitons from a data.frame of \code{frequencies}.
}
\examples{
f1 <- data.frame( Population = "A",
Locus = rep( c("Loc1","Loc2","Loc3"), each=2 ),
Allele = rep( c("01","02"), times = 3 ), 
Frequency = c( 0.2, 0.8, 0.3,0.7,0.4,0.6) )
f2 <- data.frame( Population = "B",
                  Locus = rep( c("Loc1","Loc2","Loc3"), each=2 ),
                  Allele = rep( c("01","02"), times = 3 ), 
                  Frequency = c( 0.3,0.7, 0.2, 0.8,0.4,0.6 ) ) 
f3 <- data.frame( Population = "C",
                  Locus = rep( c("Loc1","Loc2","Loc3"), each=2 ),
                  Allele = rep( c("01","02"), times = 3 ), 
                  Frequency = c( 0.4,0.6, 0.3,0.7, 0.2, 0.8 ) )
rbind(f1,f2,f3) |> 
  mutate( Population = factor( Population, ordered=TRUE ) ) -> freqs 
make_populations( freqs, N = 20, F = 0.2 ) |> summary() 
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
