% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_mating.R
\name{mixed_mating}
\alias{mixed_mating}
\title{Produce the next generation but allow for some level of selfing.}
\usage{
mixed_mating(data, N = 1, s = 0)
}
\arguments{
\item{data}{A \code{data.frame} containing individuals with objects of type \code{locus}}

\item{N}{The total number of individuals in the next generation to be produced by the mating
of each individual.  If empty, it will default to 1 offspring per pair.}

\item{s}{The fraction of adults you want to be selected to produce selfed offspring, the 
default is 0 and this will produce a randomly mating population just like if you called
\code{mate()}.}
}
\value{
A \code{data.frame} that has new individuals that are the result of mixed mating
}
\description{
This function is a bit of a wrapper around the \code{mate} function in that
 it allows you to specify the inbreeding parameter F (the fraction of individuals
 that produce offspring via selfing).
}
\examples{
data(arapat)
next_gen <- mixed_mating(arapat[1:10, ], N = 1, s = 0.2)
nrow(next_gen)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
