% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilocus_diversity.R
\name{multilocus_diversity}
\alias{multilocus_diversity}
\title{Returns multilocus diversity}
\usage{
multilocus_diversity(x)
}
\arguments{
\item{x}{A data frame with \code{locus} objects in it.}
}
\value{
The fraction of that data set that have unique multilocus genotypes.
}
\description{
This function returns the unique population size of the population
 if all individuals were treated as multilocus individuals.
}
\examples{
data(arapat)
multilocus_diversity(arapat)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
