% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_matrix.R
\name{permute_matrix}
\alias{permute_matrix}
\title{Permutes rows and corresponding cols of a matrix.}
\usage{
permute_matrix(m)
}
\arguments{
\item{m}{A square matrix.}
}
\value{
A permutation (of both rows and columns) of the matrix
}
\description{
This is a quick permutation of the matrix.
}
\examples{
m <- matrix(1:16,nrow=4)
m
permute_matrix(m)
}
