% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_popgraph.R
\name{permute_popgraph}
\alias{permute_popgraph}
\title{Bootstraps individuals to see stability of graph topology.}
\usage{
permute_popgraph(data, groups, nboot = 50, ...)
}
\arguments{
\item{data}{The raw multivariate data as submitted to \code{popgraph}}

\item{groups}{The grouping of the data into nodes as submitted to \code{popgraph}}

\item{nboot}{The nubmer of times to bootstrap the individuals per group (default=50)}

\item{...}{Other arguments to be passed to \code{popgraph}}
}
\value{
A weighted graph where edge weights represent the proportion of times the 
 edge was found in the perumuted data sets.
}
\description{
This function uses a permutation test to look at edge stability.  What we 
 do is resample individuals and re-estimate the topology several times. This
 provides an estimate of edge stability.
}
\examples{
data(arapat)
mv <- to_mv(arapat)
\donttest{
  boot_g <- permute_popgraph(mv, arapat$Population, nboot = 5)
} 
}
