% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pies_on_map.R
\name{pies_on_map}
\alias{pies_on_map}
\title{Create a pie chart map of allele frequencies on google maps.}
\usage{
pies_on_map(x, stratum = "Population", locus = NULL,
  longitude = "Longitude", latitude = "Latitude", max.rad = NULL, ...)
}
\arguments{
\item{x}{An object of type \code{data.frame} containing \code{locus} objects.}

\item{stratum}{The stratum to use for calculating frequencies (default 'Population')}

\item{locus}{The name of the locus to use (required)}

\item{longitude}{The name of the Longitude data column (default 'Longitude')}

\item{latitude}{The name of the Latitude data column (default 'Latitude')}

\item{max.rad}{The maximum radius size for the pie charts (in real space) as meters.  By 
default, it takes the maximum distance between stratum divided by 50.  They do not change
by zooming in and out, so you can 'tune' it in here.}

\item{...}{Additional arguments to plotGoogleMaps function.}
}
\value{
Nothing
}
\description{
This function takes a data.frame and makes a set of pie charts for allele frequencies
}
\examples{
\dontrun{
data(arapat)
pies_on_map(arapat,locus="LTRS")
}
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

