\name{plot.genetic.diversity}
\alias{plot.genetic.diversity}
\title{Plots out results of genetic diversity}
\usage{
  plot.genetic.diversity(x, y, ...)
}
\arguments{
  \item{x}{A \code{genetic.diversity} object}

  \item{y}{The name of the locus to plot, if this is
  missing then all loci will be examined and you will be
  prompted to cycle through them.}

  \item{...}{Ignored}
}
\description{
  Overload of plot function for genetic.diversity
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

