\name{plot.genetic.structure}
\alias{plot.genetic.structure}
\title{Plots out results of genetic structure}
\usage{
  \method{plot}{genetic.structure} (x, y, ...)
}
\arguments{
  \item{x}{A \code{genetic.structure} object}

  \item{y}{Ignored}

  \item{...}{Ignored}
}
\description{
  Overload of plot function for structure statistic
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

