\name{plot.genetic_distance}
\alias{plot.genetic_distance}
\title{Overload plot function}
\usage{
  \method{plot}{genetic_distance} (x, ...)
}
\arguments{
  \item{x}{A \code{genetic_distance} object}

  \item{...}{Ignored}
}
\description{
  This function is just for quick plotting of genetic
  distance objects
}
\author{
  Rodney J. Dyer \email{rjdyer@vcu.edu}
}

