\name{plot.structure_statistic}
\alias{plot.structure_statistic}
\title{Plot a genetic statistic}
\usage{
  \method{plot}{structure_statistic} (x, ...)
}
\arguments{
  \item{x}{An object of type 'genetic statistic'}

  \item{...}{Ignored}
}
\value{
  A ggplot object that is printed.
}
\description{
  This function plots out the observed and permuted values
  for the statistic
}
\author{
  Rodney J. Dyer \email{rjdyer@vcu.edu}
}

