% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_graph.R
\name{population_graph}
\alias{population_graph}
\title{Convience function to create population graph object from genotypes.}
\usage{
population_graph(x, stratum = "Population", numLoci = NULL, ...)
}
\arguments{
\item{x}{A \code{data.frame} with \code{locus} objects}

\item{stratum}{The stratum column to use as the Node designation (default="Population").}

\item{numLoci}{If not \code{NULL} then how many randomly selected loci (fewer than the total)
to use in the estimation.}

\item{...}{Other parameters passed to the \code{popgraph()} function.}
}
\value{
A \code{popgraph} object (an \code{igraph} graph).
}
\description{
This function is a convience function that wraps together the subset_loci() (potentially)
  function with the popgraph() function (and translations from genotypes to
  multivariate data).
}
\examples{
data(arapat)
g <- population_graph(arapat, stratum = "Population")
g
}
