\name{print.genetic.diversity}
\alias{print.genetic.diversity}
\title{Prints out results of genetic diversity}
\usage{
  print.genetic.diversity(x, ...)
}
\arguments{
  \item{x}{A \code{genetic.diversity} object}

  \item{...}{Ignored}
}
\description{
  Overload of print function for genetic.diversity
  statistic
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

