\name{print.genetic.structure}
\alias{print.genetic.structure}
\title{Prints out results of genetic structure}
\usage{
  print.genetic.structure(x, ...)
}
\arguments{
  \item{x}{A \code{genetic.structure} object}

  \item{...}{Ignored}
}
\description{
  Overload of print function for structure statistic
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

