\name{print.paternity}
\alias{print.paternity}
\title{Print paternity object}
\usage{
  \method{print}{paternity} (x, ...)
}
\arguments{
  \item{x}{An object of type class(x) = "paternity}

  \item{...}{Ignored}
}
\description{
  Overload function for printing out a paternity result for
  a single family
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

