\name{print.structure_statistic}
\alias{print.structure_statistic}
\title{Print a genetic statistic}
\usage{
  \method{print}{structure_statistic} (x, ...)
}
\arguments{
  \item{x}{An object of type 'genetic statistic'}

  \item{...}{Ignored}
}
\value{
  The text
}
\description{
  This function prints out the observed and permuted values
  for the statistic
}
\author{
  Rodney J. Dyer \email{rjdyer@vcu.edu}
}

