\name{print.summary.gstudio}
\alias{print.summary.gstudio}
\title{Printing of a summary.Locus object}
\usage{
  \method{print}{summary.gstudio} (x, ...)
}
\arguments{
  \item{x}{A \code{summary.gstudio} object}

  \item{...}{Ignored for the most part.}
}
\description{
  This prints the \code{summary.gstudio} S3 class.  Several
  of the objects found in the \code{gstudio} package are
  summarized as \code{list} objcts. This prints out their
  content in a regular fashion.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

