% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_extract.R
\name{raster_extract}
\alias{raster_extract}
\title{Extracts values from a raster at Points}
\usage{
raster_extract(
  x,
  raster,
  Latitude = "Latitude",
  Longitude = "Longitude",
  CRS = 4326
)
}
\arguments{
\item{x}{A data frame object.}

\item{raster}{A raster from which you can extract the values}

\item{Latitude}{The name of the column holding the latitude values (default = "Latitude")}

\item{Longitude}{The name of the column holding the longitude (default = "Longitude")}

\item{CRS}{The coordinate reference system (default = 4326)}
}
\description{
This is s quick convenience function that takes a `data.frame` and
  a raster and returns the extracted values as a single column.
}
