\name{read.pgraph}
\alias{read.pgraph}
\title{Loads a population graph file *.pgraph}
\usage{
  read.pgraph(file)
}
\arguments{
  \item{file}{The full path to the pgraph file}
}
\value{
  A fully formed \code{igraph} object representing the
  population graph
}
\description{
  General function to load a population graph file and turn
  it into a population graph object.
}
\author{
  Rodney J. Dyer <rjdyer@vcu.edu>
}

