\name{rel_queller}
\alias{rel_queller}
\title{This estimates the Queller & Goodnight (1989)}
\usage{
rel_queller(x, allele = NA)
}
\arguments{
  \item{x}{Either a}
}
\value{
A matrix of pairwise values for either the allele or
}
\description{
This function estimates the allele-wise relatedness
statistic from Queller & Goodnight (1989).
}
\note{
For missing data and for data comparing heterozygotes at a
2-allele locus this estimator is undefined.
}

