% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rel_ritland}
\alias{rel_ritland}
\title{This estimates the relatedness statistic from Ritland (1996)}
\usage{
rel_ritland(x)
}
\arguments{
\item{x}{A \code{data.frame} or \code{vector} that has \code{locus} objects
in it.  If you pass it a \code{data.frame} it will return the multilocus
relatedness.}
}
\value{
A matrix of relatedness statistics.
}
\description{
This function takes a \code{locus} vector or a \code{data.frame} of
 locus objects and produces a pairwise distance matrix of relatedness
 values after Ritland (1996).
}
\author{
Rodney J. dyer <rjdyer@vcu.edu>
}

