% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_coordinates.R
\name{strata_coordinates}
\alias{strata_coordinates}
\title{Grab coordinates for strata}
\usage{
strata_coordinates(
  x,
  stratum = "Population",
  longitude = "Longitude",
  latitude = "Latitude",
  sort.output = FALSE,
  single.stratum = TRUE
)
}
\arguments{
\item{x}{A \code{data.frame} object.}

\item{stratum}{The name of the stratum to partition on (default="Population").}

\item{longitude}{The column name of the longitude (default="Longitude").}

\item{latitude}{The column name of the latitude (default="Latitude").}

\item{sort.output}{A flag indicating if the results should be sorted alphabetically (default=FALSE).}

\item{single.stratum}{A flag to indicate that you only want one entry per stratum (for collapsing
points within strata, default=TRUE).}
}
\value{
A data frame with Stratum, Longitude, and Latitude, summarized by center of each stratum.
}
\description{
This function takes a \code{data.frame}, and a stratum and makes a data frame
 consisting of Stratum, Latitude, and Longitude for each stratum.
}
\examples{
data(arapat)
coords <- strata_coordinates(arapat)
head(coords)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
