% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_distance.R
\name{strata_distance}
\alias{strata_distance}
\title{Return distances between strats}
\usage{
strata_distance(x, mode = c("Euclidean", "Circle")[2])
}
\arguments{
\item{x}{A \code{data.frame} object from \code{strata_coordinates}.}

\item{mode}{The type of distance to calculate.  Currently there are two 
available types of physical distance, Euclidean (straight-line) and
Great Circle (from the curvature of the earth)}
}
\value{
A matrix of pairwise distances among strata.
}
\description{
This function takes a \code{data.frame} from the \code{strata_coordinates}
 function and returns the between strata distance matrix.
}
\note{
This uses the mean radius of the earth to be 6371km as the latest estimate by USGS.
}
\examples{
data(arapat)
coords <- strata_coordinates(arapat)
D_geo <- strata_distance(coords, mode = "Circle")
D_geo[1:5, 1:5]
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
