% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_loci.R
\name{subsample_loci}
\alias{subsample_loci}
\title{Generate subset of data with random sample of loci.}
\usage{
subsample_loci(x, numLoci)
}
\arguments{
\item{x}{A \code{data.frame} that has \code{Locus} objects in it.}

\item{numLoci}{The number of loci you wnat to sample (must be \code{length(column_class("locus"))})}
}
\value{
A new \code{data.frame} with a subset of loci.
}
\description{
This function is to assist in estimating statistical strength of your
  inferences by subsampling the loci you have.
}
\examples{
data(arapat)
sub <- subsample_loci( arapat, numLoci = 4 )
summary( sub )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
