% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_congruence.R
\name{test_congruence}
\alias{test_congruence}
\title{Returns distance congruence between two graphs}
\usage{
test_congruence(graph1, graph2)
}
\arguments{
\item{graph1}{An object of type \code{igraph} or \code{popgraph}.}

\item{graph2}{An object of type \code{igraph} or \code{popgraph}.}
}
\value{
An \code{htest} object as returned by \code{cor.test()}.
}
\description{
Computes the shortest-path distance matrices for both graphs and returns the
result of a Pearson correlation test between pairwise distances. Only nodes
present in both graphs are used.
}
\examples{
data(lopho)
test_congruence(lopho, lopho)
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
