% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_dfdist.R
\name{to_dfdist}
\alias{to_dfdist}
\title{Convert to DFDist format}
\usage{
to_dfdist(x, stratum = "Population", verbose = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} with some \code{locus} objects in it.}

\item{stratum}{The name of the 'Population' column in the data 
(default="Population").}

\item{verbose}{A flag to print locus names as you go (default=FALSE)}
}
\value{
A text object with data formatted for output.
}
\description{
This converts a \code{data.frame} object with loci into a 
  format suitable for FDist, DFDist, & DetSel
}
\examples{
data(arapat)
txt <- to_dfdist(arapat[1:30, ], stratum = "Population")
cat(substr(txt, 1, 100))
}
