% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_fixed_locus.R
\name{to_fixed_locus}
\alias{to_fixed_locus}
\title{returns string representation of locus for output like genepop/heirfstat}
\usage{
to_fixed_locus(x, digits = NULL)
}
\arguments{
\item{x}{An object of type \code{locus}.}

\item{digits}{The number of digits that the alleles need to have}
}
\value{
A character representation of the locus, one element per entry of \code{x}.
}
\description{
This function
}
\examples{
loc <- locus(c(1, 12))
to_fixed_locus(loc, digits = 2)
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
