% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_geojson.R
\name{to_geojson}
\alias{to_geojson}
\title{Converts popgraph to geojson for import into ArcGIS Online}
\usage{
to_geojson(graph, file)
}
\arguments{
\item{graph}{A popgraph object.}

\item{file}{The file for saving the GeoJSON}
}
\value{
A textual version of the popgraph
}
\description{
This is a convienence function to convert a population graph
  object into GeoJSON format for inclusion into ArcGIS Online
}
\examples{
data(lopho)
data(baja)
graph <- decorate_graph(lopho, baja)
json <- to_geojson(graph)
cat(substr(json, 1, 100))
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
