% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json.R
\name{to_json}
\alias{to_json}
\title{Converts to json representation}
\usage{
to_json(graph, file, forJS = FALSE)
}
\arguments{
\item{graph}{An object of type \code{igraph}}

\item{file}{A file to write the output to (default missing)}

\item{forJS}{A flag to indicate putting a 'var myjson = ' part 
the file.}
}
\value{
A textual json representation of the graph
}
\description{
This is a nice function that takes a graph structure
 and converts it to a json format for use on the web.
}
\examples{
data(lopho)
json <- to_json(lopho)
cat(substr(json, 1, 100))
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
