% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sf.R
\name{to_sf}
\alias{to_sf}
\title{This function convers the popgraph object into an object that can be plot using
 the leaflet library.}
\usage{
to_sf(
  x,
  what = c("nodes", "edges")[1],
  Longitude = "Longitude",
  Latitude = "Latitude",
  CRS = 4326
)
}
\arguments{
\item{x}{A population graph}

\item{what}{An indication of either \code{node} or \code{edge} objects to be returned
for plotting on a leaflet object.}

\item{Longitude}{The name of the column with the longitude (or equivalent) coordinates}

\item{Latitude}{The name of the column with the latitude (or equivalent) coordinate.}

\item{CRS}{The coordinate reference system (see https://epsg.io for values).}
}
\value{
A \code{data.frame} object suitable for passing to \code{leaflet}.
}
\description{
This function convers the popgraph object into an object that can be plot using
 the leaflet library.
}
\examples{
# library(igraph)
# data(lopho)
# data(baja)
# graph <- decorate_graph(lopho, baja )
# nodes <- to_sf( graph, what="nodes")
# edges <- to_sf( graph, what="edges")
# library( leaflet )
# leaflet() |>
#  addTiles() |>
#  addMarkers( data=nodes ) |>
#  addPolylines( data=edges )
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
