% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_SpatialLines.R
\name{to_SpatialLines}
\alias{to_SpatialLines}
\title{Convert popgraph edges to a SpatialLines object}
\usage{
to_SpatialLines(graph, latitude = "Latitude", longitude = "Longitude", ...)
}
\arguments{
\item{graph}{A \code{popgraph} object decorated with \code{Latitude} and
\code{Longitude} vertex attributes (see \code{\link{decorate_graph}}).}

\item{latitude}{Name of the vertex attribute holding latitude values
(default \code{"Latitude"}).}

\item{longitude}{Name of the vertex attribute holding longitude values
(default \code{"Longitude"}).}

\item{...}{Ignored.}
}
\value{
A \code{sp::SpatialLines} object.
}
\description{
`r lifecycle::badge("deprecated")`

\code{to_SpatialLines()} is deprecated. The \code{sp} package is being
superseded by \code{sf}. Use \code{\link{to_sf}} instead, which returns an
\code{sf} object compatible with modern spatial workflows.
}
\seealso{
\code{\link{to_sf}}
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
