% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_counts.R
\name{allele_counts}
\alias{allele_counts}
\title{Returns counts of alleles at a locus by population}
\usage{
allele_counts(x, locus, stratum = "Population")
}
\arguments{
\item{x}{A \code{data.frame} object with loci and stratum in it.}

\item{locus}{The name of the locus to use}

\item{stratum}{The stratum to partition by.}
}
\value{
A matrix where each row is a strata and each column is a count of alleles
}
\description{
Takes a locus and set of populations and returns matrix of
  allele counts
}
\author{
Rodney J. Dyer
}

