% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_data_frame.R
\name{as.data.frame.genind}
\alias{as.data.frame.genind}
\title{Converts genind to normal data.frame with \code{locus} objects}
\usage{
\method{as.data.frame}{genind}(x, row.names, optional, ...)
}
\arguments{
\item{x}{An object of type \code{genind}}

\item{row.names}{ignored here}

\item{optional}{ignored here}

\item{...}{ignored here}
}
\value{
A \code{data.frame} object with loci represented as \code{locus} objects.
}
\description{
This is a conversion function to translate an object of type \code{genind} from
 the adegenet package into a normal data frame with \code{locus} objects.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

