% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_popgraph.R
\name{as.popgraph}
\alias{as.popgraph}
\title{Converts an object of type matrix or igraph into \code{popgraph}}
\usage{
as.popgraph(graph)
}
\arguments{
\item{graph}{An object of type \code{matrix} or \code{igraph}}
}
\value{
An object of type \code{popgraph}
}
\description{
This is a simple conversion routine for \code{matrix} or \code{igraph} 
 objects into \code{popgraph} objects
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
