% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetric_weights.R
\name{asymmetric_weights}
\alias{asymmetric_weights}
\title{Asymmetric Edge Weights
 This uses a method inspired by t-distributed Stochastic Neighbor Embedding
 to estimate relative weights of edges in a population graph using local 
 embedding and Markov properties.}
\usage{
asymmetric_weights(graph, perplexity = 4, tol = 1e-05, max_iter = 100)
}
\arguments{
\item{graph}{An existing ``popgraph`` object.}

\item{perplexity}{A target perplexity value for the bandwidth search (default 4).}

\item{tol}{A minimal tolerance for variance values}

\item{max_iter}{An upper limit on the iterations necessary for finding a solution.}
}
\value{
A directional population graph.
}
\description{
Asymmetric Edge Weights
 This uses a method inspired by t-distributed Stochastic Neighbor Embedding
 to estimate relative weights of edges in a population graph using local 
 embedding and Markov properties.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
