% Generated by roxygen2 (4.0.1): do not edit by hand
\name{c.locus}
\alias{c.locus}
\title{Concatinate \code{locus} objects}
\usage{
\method{c}{locus}(..., recursive = FALSE)
}
\arguments{
\item{\dots}{The \code{locus} objects to be concatenated.}

\item{recursive}{A flag to do the concatenation recursively.}
}
\value{
A vector of \code{locus} objects
}
\description{
An overload of the \code{c} function for \code{locus} objects.
}
\examples{
loci <- c( locus(1:2), locus(1:2), locus( c(1,1) ) )
loci
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

