% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2dist.R
\name{cov2dist}
\alias{cov2dist}
\title{Converts covariance matrix to distance one.}
\usage{
cov2dist(C)
}
\arguments{
\item{C}{A pairwise, square, and symmetric distance matrix.}
}
\value{
A distance matrix of the same size.
}
\description{
This function takes a covariance matrix and turns it into a distance
  matrix following Gower (1966).
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
