% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_to_covariance.R
\name{distance_to_covariance}
\alias{distance_to_covariance}
\title{Takes distance matrix and turns it into a covariance matrix}
\usage{
distance_to_covariance(X)
}
\arguments{
\item{X}{A square and symmetric distance matrix.}
}
\value{
A covariance matrix the same size as \code{X}
}
\description{
This function takes a distance matrix and turns it into a 
  covariance matrix.
}

