% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_loci.R
\name{.mutate_allele}
\alias{.mutate_allele}
\title{Mutate a single allele}
\usage{
.mutate_allele(allele, mutation, all_numeric)
}
\arguments{
\item{allele}{Character string of the allele.}

\item{mutation}{A \code{mutation_model} object.}

\item{all_numeric}{Numeric vector of all alleles seen (for IAM).}
}
\value{
Character string of the new allele, or NULL if skipped.
}
\description{
Mutate a single allele
}
\keyword{internal}
