% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_contortion.R
\name{edge_contortion}
\alias{edge_contortion}
\title{Determines of edges are stretched or compressed and adds attribute
 to popgraph object.}
\usage{
edge_contortion(
  graph,
  latitude = "Latitude",
  longitude = "Longitude",
  P = NULL
)
}
\arguments{
\item{graph}{A \code{popgraph} object}

\item{latitude}{The name of the node property that has the value of the latitude for
that node (default 'Latitude').}

\item{longitude}{The name of the edge property that has the longitude (default 'Longitude')}

\item{P}{A non-genetic distance (physical or ecological) matrix on which the genetic distances 
are compared.}
}
\value{
A copy of the graph with new properties 'contortion' (the magnitude of differences
  between genetic and physical distances) and 'stretch' (as 'Compressed' or 'Extended')
}
\description{
This function measure the deviation of edges in physical space from the expected 
  value given by the genetic distance.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
