% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_matrix.R
\name{frequency_matrix}
\alias{frequency_matrix}
\title{Create a matrix of allele frequencies for a single locus across many strata}
\usage{
frequency_matrix(x, stratum = "Population", loci = NULL)
}
\arguments{
\item{x}{A \code{data.frame} with locus and strata column(s)}

\item{stratum}{The column name to use as the partitioning}

\item{locus}{The name of the locus to use}
}
\value{
A matrix (rows are strata, columns are alleles) of allele frequencies
}
\description{
This is a quick translation of locus data to a tabular stratum format.
}
\note{
This function just reshapes the \code{data.frame} from \code{frequencies}.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
