% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_autocorrealtion.R
\name{genetic_autocorrelation}
\alias{genetic_autocorrelation}
\title{Performs 'Autocorrelation Analysis}
\usage{
genetic_autocorrelation(P, G, bins, mode = c("smouse", "moran")[1])
}
\arguments{
\item{P}{A square physical distance matrix}

\item{G}{A square genetic distance matrix (same size as P)}

\item{binsize}{The distance bins to be estimated.  These will be taken sequentially 
such that the first bin will be all observations whose separation is greater than
binsize[1] and less than binsize[2].}
}
\description{
This function takes two distance matices, one for physical separation
 and the other for genetic separation as well as a bin size and performs
 the spatial autocorrelation analysis of Smouse & Peakall (1999).
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
