% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_distance.R
\name{genetic_distance}
\alias{genetic_distance}
\title{Estimate genetic distances among individuals or strata}
\usage{
genetic_distance(x, stratum = "Population", mode)
}
\arguments{
\item{x}{A \code{data.frame} object with \code{\link{locus}} columns, or a single \code{locus} vector.}

\item{stratum}{The strata by which genetic distances are estimated (default="Population").
Optional for individual-based distance metrics like AMOVA and Bray.}

\item{mode}{The genetic distance metric to use:
\describe{
  \item{amova}{Inter-individual squared Euclidean distance (Peakall et al. 1995, Smouse & Peakall 1999)}
  \item{bray}{Inter-individual Bray-Curtis / shared allele distance (Bowcock et al. 1994)}
  \item{dps}{Stratum-level shared allele distance (1 - Ps)}
  \item{cavalli}{Cavalli-Sforza & Edwards (1967) chord distance}
  \item{cgd}{Conditional Genetic Distance via Population Graphs (Dyer & Nason 2004)}
  \item{euclidean}{Euclidean allele frequency distance}
  \item{jaccard}{Jaccard set dissimilarity}
  \item{nei}{Nei's unbiased genetic distance (1978)}
  \item{ss}{Partitioned Sum of Squares distance from AMOVA}
}}
}
\value{
A pairwise distance matrix.
}
\description{
This is the primary front-end function for estimating genetic distances among
 either individuals or strata.
}
\examples{
AA <- locus( c("A","A") )
AB <- locus( c("A","B") )
BB <- locus( c("B","B") )
loci <- c(AA,AB,AB,BB,BB,AA,AB,BB,AA,BB) 
df <- data.frame( Population=c(rep("Pop-A",5), rep("Pop-B",5)), TPI=loci)
genetic_distance(df, mode="amova")
genetic_distance(df, mode="dps")
genetic_distance(df, mode="cavalli")
genetic_distance(df, mode="nei")
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
