\name{geom_surface}
\alias{geom_surface}
\title{Translate raster into ggplot object}
\usage{
geom_surface(raster, ...)
}
\arguments{
  \item{raster}{An object of type raster}

  \item{...}{Passed on to \code{geom_tile} as optional
  arguments.}
}
\value{
An object of type geom_tile
}
\description{
This function takes a raster and translates it into a
ggplot geometry layer for plotting.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

