\docType{package}
\name{gstudio-package}
\alias{gstudio-package}
\title{Routines used in spatial analysis of genetic marker data}
\description{
gstudio is a suite of tools used in the spatial analysis of
population genetic data created for the book "Applied
Landscape Genetics" by R.J. Dyer.  This version of the
package (1.X) is a significant deviant from the previous
(versions 0.X) resulting from an entire re-write of the
package to maximize ease of use and algorithm quickness.
As such, previous scripts using gstudio (vers. 0.X) will
need to be refactored a bit.  While this can be a bit of a
pain, the larger utility of the new format will pay off in
spades.
}
\details{
\tabular{ll}{ Package: \tab gstudio\cr Type: \tab
Package\cr Version: \tab 1.1\cr Date: \tab 2013-06-07\cr
License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }

There are some very good examples of the components of this
package are used in the vignettes for this package.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
\keyword{package}

