% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/he.R
\name{He}
\alias{He}
\title{Estimate expected heterozygosity}
\usage{
He(x, small.N = FALSE)
}
\arguments{
\item{x}{Either a \code{data.frame} object with \code{locus} objects or
a vector or \code{locus} objects.}

\item{small.N}{Apply the 2N/(2N-1) correction to the data
for small sample sizes.}
}
\value{
The expected heterozygosity as a numeric or a \code{data.frame} if
 several loci are passed.
}
\description{
Returns the general expected heterozygosity parameter
 from the frequencies
}
\examples{
data(arapat)
He(arapat)
He(arapat, small.N = TRUE)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
