% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ho.R
\name{Ho}
\alias{Ho}
\title{Estimate observed heterozygosity
 
Returns the general observed heterozygosity parameter
 from the frequencies}
\usage{
Ho(x)
}
\arguments{
\item{x}{An object of type \code{locus}}
}
\value{
The expected heterozygosity
}
\description{
Estimate observed heterozygosity
 
Returns the general observed heterozygosity parameter
 from the frequencies
}
\note{
This function can be called on a single vector of data, a \code{data.frame} of loci, 
 or a \code{data.frame} of \code{locus} objects across strata.  If the estimating across
 stratum, the unbiased estimator should be used to average across stratum and is performed
 by passing the appropriate stratum= argument.
}
\examples{
loci <- c( locus( c("A","A") ), locus( c("A","A") ), locus( c("A","B") ) )
Ho( loci )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
