% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwe.R
\name{hwe}
\alias{hwe}
\title{Tests for Hardy-Weinberg Equilibrium}
\usage{
hwe(x, mode = c("Chi")[1], supress_warnings = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} with one or more \code{locus} objects in it}

\item{mode}{The way in which the probabilities are estimated.  Possible values include
'Chi': Chi-square approximation & 'Permute': A permutation approach}

\item{supress_warnings}{A flag that prevents sample size warnings from being displayed
(often helpful if you are simulating data sets for permutation)}
}
\value{
A \code{data.frame} with columns for Locus, Chi (the stat), df, and Prob.
}
\description{
This function tests for Hardy-Weinberg Equilibrium using the either 
 the chi-square approximation or a permutation approach.
}
\examples{
 data(arapat)
 sonora <- arapat[ arapat$Species=="Mainland",]
 hwe( sonora )
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

