\name{is_frequency}
\alias{is_frequency}
\title{Is-A function for Allele Frequences}
\usage{
is_frequency(x)
}
\arguments{
  \item{x}{A \code{data.frame} potentially from
  \code{frequencies} function.}
}
\value{
A flag indicating it has the qualities of a
\code{data.frame} representing allele frequencies
}
\description{
This is a function that returns a flag indicating that the
object passed has the qualities of a \code{data.frame}
created by the function \code{frequencies}.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

