% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_matrix.R
\name{lag_matrix}
\alias{lag_matrix}
\title{Returns matrix only with values within a particular bin distance range}
\usage{
lag_matrix(X, lower = NULL, upper = NULL)
}
\arguments{
\item{X}{The pairwise distance matrix.}

\item{lower}{The lower limit of the bin being requested}

\item{upper}{The upper limit of the bin being requested}
}
\value{
A matrix with the count of objects in each bin category.
}
\description{
This function returns bin distances for spatial autocorrelation.
}
