% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstudio-package.R
\docType{data}
\name{lopho}
\alias{lopho}
\title{Lophocereus population graph}
\description{
This is the population graph for the Lophocereus data
 that is discussed in Dyer & Nason (2004).
}
\keyword{data}
