% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_allele_size.R
\name{maximum_allele_size}
\alias{maximum_allele_size}
\title{Returns the max digits of all loci passed.}
\usage{
maximum_allele_size(x)
}
\arguments{
\item{x}{A \code{data.frame} or vector of \code{locus} objects}
}
\value{
An integer depicting the maximum number of digits in alleles at the 
 locus.
}
\description{
This is a convienence function that returns the number of digits for a locus
 or set of loci.
}
\examples{
x <- c( locus(c(1,2)), locus(c(121,23)))
maximum_allele_size( x )
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

