% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_ci.R
\name{permute_ci}
\alias{permute_ci}
\title{Permute for confidence interval}
\usage{
permute_ci(
  x,
  stratum = NULL,
  allele_perm = TRUE,
  nperm = 99,
  FUN = NULL,
  replace = FALSE,
  ...
)
}
\arguments{
\item{x}{The \code{locus} data being used.}

\item{stratum}{An optional argument if passed, permutes \code{x} among 
strata and gives the results to the function.}

\item{allele_perm}{A flag (default=TRUE) to permute alleles in the
creating the new data set.}

\item{nperm}{The number of perms to use (default 99).}

\item{FUN}{The function to pass the permuted data to.  It is called on a
\code{locus} vector (and the permuted strata, if given) and must return a
single number; wrap \code{\link{genetic_diversity}} as in the example to
permute a diversity statistic.}

\item{replace}{A flag indicating if randomization should be with replacement (default FALSE)}

\item{...}{Passed on as arguments to \code{FUN}.}
}
\value{
A vector of permuted response values from the function.
}
\description{
This function takes some data and does permutations on it and passes
 it along to the indicated function.  There are options for permuting
 genotypes or permuting alleles.  Make sure you know the difference.
}
\examples{
loc <- c(locus(1:2), locus(c(1,1)), locus(c(2,2)))
he <- function(x) genetic_diversity(x, mode = "He")$He
permute_ci(loc, FUN = he, nperm = 19)
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
