\name{plot.locus}
\alias{plot.locus}
\title{Overload plot function}
\usage{
  \method{plot}{locus} (x, mode = c("bar", "pie")[1], ...)
}
\arguments{
  \item{x}{A set of \code{locus} objects}

  \item{mode}{The type of plot to create, bar or pie (bar
  is default).}

  \item{...}{Ignored}
}
\description{
  This function is just for quick plotting of frequencies
}
\author{
  Rodney J. Dyer \email{rjdyer@vcu.edu}
}

