% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_graph.R
\name{randomize_graph}
\alias{randomize_graph}
\title{Randomize Graph}
\usage{
randomize_graph(graph = NULL, mode = c("full", "degree")[2])
}
\arguments{
\item{graph}{An object of type \code{popgraph} or \code{igraph}}

\item{mode}{The kind of randomization to conduct, can be "full"
which makes a new graph with the same number of edges as the 
original one, or "degree" which preserves the degree distribution
of the}
}
\description{
This function randomizes the edges in a popgraph and 
 returns a new graph.
}
