% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon Entry Function
 The degree of uncertainty in a system based on probability and information theory.}
\usage{
shannon_entropy(p)
}
\arguments{
\item{p}{The probility}
}
\value{
An estimate of entropy
}
\description{
Shannon Entry Function
 The degree of uncertainty in a system based on probability and information theory.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
