% Generated by roxygen2 (4.0.1): do not edit by hand
\name{strata_coordinates}
\alias{strata_coordinates}
\title{Grab coordinates for strata}
\usage{
strata_coordinates(x, stratum = "Population", longitude = "Longitude",
  latitude = "Latitude", as.SpatialPoints = FALSE, sort.output = FALSE)
}
\arguments{
\item{x}{A \code{data.frame} object.}

\item{stratum}{The name of the stratum to partition on (default="Population").}

\item{longitude}{The column name of the longitude}

\item{latitude}{The column name of the latitude}

\item{as.SpatialPoints}{A flag indicating what kind of coordinates to return
should be turned into a SpatialPoints object (TRUE) or as a \code{data.frame} (FALSE,
the default)}

\item{sort.output}{A flag indicating if the results should be sorted alphabetically (default=FALSE)}
}
\value{
A data frame, with Stratum Latitude and Longitude, summarized by center of each stratum.
}
\description{
This function takes a , and a stratum and makes a data frame
 consisting of Stratum, Latitide, and Longitude for each stratum
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}

