% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_congruence.R
\name{test_congruence}
\alias{test_congruence}
\title{Returns distance congruence between the two graphs}
\usage{
test_congruence(graph1, graph2, method = c("distance", "combinatorial")[1])
}
\arguments{
\item{graph1}{An object of type \code{igraph} or \code{popgraph}}

\item{graph2}{An object of type \code{igraph} or \code{popgraph}}

\item{method}{An option on how congruence is to be estimated
possible values are 'distance' (a measure of similarity in 
separation of nodes independent of connectivity, the default) 
'structural' a measure of similarity in actual edges, and 
'combinatorial' a combinatorial measure of similarity.}
}
\value{
A non-parametric rank sum test
}
\description{
This function makes the shortest path matrices for both
 graphs and determines the correlation between pairwise
 distance.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
