% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data.frame.R
\name{to_data.frame}
\alias{to_data.frame}
\title{Converts \code{popgraph} to \code{data.frame} based upon node attributes}
\usage{
to_data.frame(x, mode = c("nodes", "edges")[1], as.named = TRUE, ...)
}
\arguments{
\item{x}{The \code{popgraph} to grab stuff from.}

\item{mode}{An indication of what to return, a \code{data.frame} of node (the default) 
or edge properties.}

\item{as.named}{A flag (default \code{TRUE}) indicating whether the nodes are returned
as indices or names.}

\item{...}{Ignored (generally).}
}
\value{
An object of type \code{data.frame} with all the node attributes.
}
\description{
This is a quick conversion of vertex attributes to a \code{data.frame}, essentially
 the reverse operation as \code{decorate_graph} function.
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}
}
