% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df}
\alias{to_df}
\title{Converts popgraph into Data.Frame}
\usage{
to_df(graph, mode = c("nodes", "edges")[1])
}
\arguments{
\item{graph}{An population graph}

\item{mode}{Either "nodes" or "edges" to be used.}
}
\value{
A \code{data.frame}.  For \code{mode = "nodes"}, one row per node
  with the node name as \code{Stratum} followed by the other vertex
  attributes; for \code{mode = "edges"}, one row per edge with \code{from},
  \code{to} and the edge attributes.
}
\description{
Quick function to convert a population graph into a data.frame
  object for either the node or edge set with all attributes.
}
\examples{
data(lopho)
df_edges <- to_df(lopho, mode = "edges")
head(df_edges)

}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
