% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_df.R
\name{to_df}
\alias{to_df}
\title{Converts popgraph into Data.Frame}
\usage{
to_df(graph, mode = c("nodes", "edges")[1])
}
\arguments{
\item{graph}{An population graph}

\item{mode}{Either "nodes" or "edges" to be used.}
}
\value{
An object of type data.frame
}
\description{
Quick function to convert a population graph into a data.frame
  object for either the node or edge set with all attributes.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
