\name{to_fixed_locus}
\alias{to_fixed_locus}
\title{returns string representation of locus for output like genepop/heirfstat}
\usage{
to_fixed_locus(x, digits = NULL)
}
\arguments{
  \item{x}{An object of type \code{locus}.}

  \item{digits}{The number of digits that the alleles need
  to have}
}
\value{
A character representation of the locus with
}
\description{
This function
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}

