% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_mv.R
\name{to_mv}
\alias{to_mv}
\alias{to_mv.default}
\alias{to_mv.locus}
\alias{to_mv.data.frame}
\title{Returns the passed items as multivariate data}
\usage{
to_mv(x, ploidy, alleles, drop.allele = FALSE, leave.as.na = FALSE, ...)

\method{to_mv}{default}(x, ...)

\method{to_mv}{locus}(x, ploidy = 2, alleles = NA, drop.allele = FALSE,
  leave.as.na = FALSE, ...)

\method{to_mv}{data.frame}(x, ploidy = 2, alleles = NA,
  drop.allele = FALSE, leave.as.na = FALSE, ...)
}
\arguments{
\item{x}{A object to convert.}

\item{ploidy}{The number of alleles at an individuals locus (default=2)}

\item{alleles}{A set of alleles to make the matrix columns on.}

\item{drop.allele}{A flag indicating an allele should be dropped.}

\item{leave.as.na}{A flag to re-insert NA values into multivariate data where missing
genotypes were encountered (Default=FALSE) if not, zeros are used.}

\item{...}{Ignored}
}
\value{
The matrix representation of \code{x}.

Numerical value of x

Matrix representing the locus

A matrix
}
\description{
This function will provide the alleles within a \code{locus} object 
 for either a single locus or for a vector of loci
}
\author{
Rodney J. Dyer \email{rjdyer@vcu.edu}

Rodney J. Dyer \email{rjdyer@vcu.edu}
}
