\name{write_population}
\alias{write_population}
\title{Writes data frame with genotypes and other data to a file}
\usage{
write_population(df, file, mode = c("Text", "GENEPOP", "STRUCTURE"))
}
\arguments{
  \item{df}{The \code{data.frame} to be written to the
  output.}

  \item{file}{The path to the file you want to write to.}

  \item{mode}{A distinction of}
}
\description{
This is a basic output file formatter for saving files
necessary for saving output in textual formats.
}

