% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_population.R
\name{write_population}
\alias{write_population}
\title{Writes data frame with genotypes and other data to a file}
\usage{
write_population(df, file, mode = c("text", "genepop", "structure",
  "dfdist")[1], stratum = NULL, ...)
}
\arguments{
\item{df}{The \code{data.frame} to be written to the output.}

\item{file}{The path to the file you want to write to.}

\item{mode}{Which format to use for writing data to file.  At present there
 are 
\describe{
 \item{text}{Outputs all data columns with loci treated as colon separated.  This is the default.}
 
 \item{genepop}{Saves genetic data into GENEPOP format. You must specify a stratum.}
 \item{structure}{Saves genetic data into the two-lined per individual STRUCTURE format.  You must specify a stratum.}
 \item{dfdist}{Saves genetic data as a FDist, DFDist, DetSel file.  You must specify the stratum.}
}}

\item{stratum}{An optional argument if using genepop or structure formats.  By 
default, the 'text' option writes all data to file.}

\item{...}{Other options to pass along (e.g., verbose=TRUE)}
}
\value{
nothing.
}
\description{
This is a basic output file formatter for saving files necessary for saving
 output in textual formats.
}
\author{
Rodney J. Dyer <rjdyer@vcu.edu>
}
